feat: ssz support for block
